fix(deps): resolve rustls-webpki and sqlx RustSec advisories
cargo audit flagged rustls-webpki 0.101.7 (RUSTSEC-2026-0104/-0099/-0098,
reachable panic + name-constraint bypasses) and sqlx 0.7.4
(RUSTSEC-2024-0363, binary-protocol truncating/overflowing casts).

rustls-webpki could not be bumped directly: it is transitive via
reqwest -> rustls 0.21.12, which pins rustls-webpki to caret 0.101.7.
Bumped the direct dependency reqwest 0.11 -> 0.12, pulling in rustls
0.23.41 -> rustls-webpki 0.103.13 (patches all three advisories).
Source usage of reqwest is a single minimal Client::builder/post/json/
send call (src/consensus/client.rs), unaffected by the 0.11->0.12 API
surface.

Bumped sqlx 0.7 -> 0.8.6 (latest 0.8.x, staying within the minor line
requested rather than jumping to the newly-released 0.9.0). This is a
binary crate whose .sqlx/ offline query cache was generated today
against sqlx 0.7.4 (commit 5ae2237) for Dockerfile.orchestrator's
SQLX_OFFLINE=true release build. Regenerated the cache for 0.8.6 via
sqlx-cli 0.8.6 (upgraded from 0.7.4 to match) against a scratch SQLite
DB with migrations applied, then deleted the scratch DB. 4 of 9 cached
queries changed: sqlx-sqlite 0.8.6 reports SQLite INTEGER columns as a
unified Integer type_info instead of the old ambiguous Int/Int64
split, which is the type-safety correction RUSTSEC-2024-0363 addresses.

That stricter typing surfaced one real, previously-latent bug: the
rate-limit count in check_rate_limit (src/api/middleware.rs) now infers
as i64 instead of i32, so comparing it directly against the i32
limit parameter no longer compiles. Fixed by widening the comparison
via i64 conversion of limit rather than truncating the count,
preserving correctness if a count ever exceeds i32 max.

Investigated rsa 0.9.10 (RUSTSEC-2023-0071, Marvin Attack timing
side-channel, no fixed version exists): it is a transitive dependency
of sqlx-mysql/sqlx-postgres, but cargo tree -i rsa and cargo metadata
resolve graph both show it as unreachable in this build -- this crate
only enables sqlx sqlite feature, never mysql or postgres. Confirmed
via cargo tree both before and after the sqlx 0.7->0.8 bump. No source
code calls rsa directly. Recommendation: no action needed; the crate
is not compiled into the binary despite appearing in Cargo.lock, so
the timing side-channel is not reachable by this codebase. If mysql or
postgres support is ever added, this must be re-evaluated at that
time.

Verified: cargo build --workspace and cargo test --workspace both
clean before (35m06s build, 82/82 tests) and after (2m58s build,
82/82 tests) -- zero regressions, same 2 pre-existing unrelated
warnings present in both. docker build -f docker/Dockerfile.orchestrator
completed end-to-end in a fresh container (25m52s release+LTO build)
using only the regenerated offline cache and SQLX_OFFLINE=true, then
the built image was run and confirmed to initialize its DB schema and
serve /api/v1/health successfully. cargo audit before: 5 vulnerabilities
(rsa, rustls-webpki x3, sqlx). cargo audit after: 1 vulnerability (rsa
only, as expected, no fix available).

+++ src/api/middleware.rs
@@ -400,7 +400,7 @@ impl RateLimitMiddleware {
 
         let current_count = count.count;
 
-        if current_count >= limit {
+        if current_count >= i64::from(limit) {
             return Err(RateLimitError::Exceeded {
                 retry_after: window_duration.as_secs(),
             });
